+++ test/plug/plug_set_locale_test.exs
@@ -12,7 +12,7 @@ defmodule Cldr.Plug.SetLocale.Test do
     ]
 
   test "init returns the default options" do
-    opts = Cldr.Plug.SetLocale.init(cldr: TestBackend.Cldr)
+    opts = Cldr.Plug.SetLocale.init(cldr_backend: TestBackend.Cldr)
 
     assert opts == [
              session_key: "cldr_locale",
@@ -35,37 +35,37 @@ defmodule Cldr.Plug.SetLocale.Test do
              param: "locale",
              from: [:session, :accept_language],
              apps: [:cldr],
-             cldr: TestBackend.Cldr
+             cldr_backend: TestBackend.Cldr
            ]
   end
 
   test "bad parameters raise exceptions" do
     assert_raise ArgumentError, fn ->
-      Cldr.Plug.SetLocale.init(from: :nothing, cldr: TestBackend.Cldr)
+      Cldr.Plug.SetLocale.init(from: :nothing, cldr_backend: TestBackend.Cldr)
     end
 
     assert_raise ArgumentError, fn ->
-      Cldr.Plug.SetLocale.init(from: :nothing, cldr: TestBackend.Cldr)
+      Cldr.Plug.SetLocale.init(from: :nothing, cldr_backend: TestBackend.Cldr)
     end
 
     assert_raise ArgumentError, fn ->
-      Cldr.Plug.SetLocale.init(from: [:nothing], cldr: TestBackend.Cldr)
+      Cldr.Plug.SetLocale.init(from: [:nothing], cldr_backend: TestBackend.Cldr)
     end
 
     assert_raise ArgumentError, fn ->
-      Cldr.Plug.SetLocale.init(apps: :nothing, cldr: TestBackend.Cldr)
+      Cldr.Plug.SetLocale.init(apps: :nothing, cldr_backend: TestBackend.Cldr)
     end
 
     assert_raise ArgumentError, fn ->
-      Cldr.Plug.SetLocale.init(apps: [:nothing], cldr: TestBackend.Cldr)
+      Cldr.Plug.SetLocale.init(apps: [:nothing], cldr_backend: TestBackend.Cldr)
     end
 
     assert_raise ArgumentError, fn ->
-      Cldr.Plug.SetLocale.init(param: [:nothing], cldr: TestBackend.Cldr)
+      Cldr.Plug.SetLocale.init(param: [:nothing], cldr_backend: TestBackend.Cldr)
     end
 
     assert_raise ArgumentError, fn ->
-      Cldr.Plug.SetLocale.init(apps: :gettext, cldr: TestBackend.Cldr)
+      Cldr.Plug.SetLocale.init(apps: :gettext, cldr_backend: TestBackend.Cldr)
     end
 
     assert_raise ArgumentError, fn ->
@@ -73,12 +73,12 @@ defmodule Cldr.Plug.SetLocale.Test do
     end
 
     assert_raise Cldr.UnknownLocaleError, fn ->
-      Cldr.Plug.SetLocale.init(default: :nothing, cldr: TestBackend.Cldr)
+      Cldr.Plug.SetLocale.init(default: :nothing, cldr_backend: TestBackend.Cldr)
     end
   end
 
   test "set the locale from a query param" do
-    opts = Cldr.Plug.SetLocale.init(from: :query, cldr: TestBackend.Cldr)
+    opts = Cldr.Plug.SetLocale.init(from: :query, cldr_backend: TestBackend.Cldr)
 
     conn =
       :get
@@ -102,11 +102,11 @@ defmodule Cldr.Plug.SetLocale.Test do
                language_variant: nil
              }
 
-    assert opts[:cldr].get_current_locale() == conn.private[:cldr_locale]
+    assert opts[:cldr_backend].get_current_locale() == conn.private[:cldr_locale]
   end
 
   test "set the locale from the session" do
-    opts = Cldr.Plug.SetLocale.init(from: :session, cldr: TestBackend.Cldr)
+    opts = Cldr.Plug.SetLocale.init(from: :session, cldr_backend: TestBackend.Cldr)
     session_opts = Plug.Session.init(store: :cookie, key: "_key", signing_salt: "X")
 
     conn =
@@ -114,7 +114,7 @@ defmodule Cldr.Plug.SetLocale.Test do
       |> conn("/?locale=fr")
       |> Plug.Session.call(session_opts)
       |> fetch_session("cldr_locale")
-      |> put_session("cldr_locale", Cldr.Locale.new("ru", opts[:cldr]))
+      |> put_session("cldr_locale", Cldr.Locale.new("ru", opts[:cldr_backend]))
       |> Cldr.Plug.SetLocale.call(opts)
 
     assert conn.private[:cldr_locale] ==
@@ -134,11 +134,11 @@ defmodule Cldr.Plug.SetLocale.Test do
                territory: "RU"
              }
 
-    assert opts[:cldr].get_current_locale() == conn.private[:cldr_locale]
+    assert opts[:cldr_backend].get_current_locale() == conn.private[:cldr_locale]
   end
 
   test "set the locale from a body param" do
-    opts = Cldr.Plug.SetLocale.init(from: :body, cldr: TestBackend.Cldr)
+    opts = Cldr.Plug.SetLocale.init(from: :body, cldr_backend: TestBackend.Cldr)
     parser_opts = Plug.Parsers.init(parsers: [:json], json_decoder: Jason)
     json = %{locale: "zh-Hant"}
 
@@ -165,11 +165,11 @@ defmodule Cldr.Plug.SetLocale.Test do
                territory: "TW"
              }
 
-    assert opts[:cldr].get_current_locale() == conn.private[:cldr_locale]
+    assert opts[:cldr_backend].get_current_locale() == conn.private[:cldr_locale]
   end
 
   test "set the locale from a cookie param" do
-    opts = Cldr.Plug.SetLocale.init(from: :cookie, cldr: TestBackend.Cldr)
+    opts = Cldr.Plug.SetLocale.init(from: :cookie, cldr_backend: TestBackend.Cldr)
 
     conn =
       :get
@@ -195,11 +195,11 @@ defmodule Cldr.Plug.SetLocale.Test do
                territory: "TW"
              }
 
-    assert opts[:cldr].get_current_locale() == conn.private[:cldr_locale]
+    assert opts[:cldr_backend].get_current_locale() == conn.private[:cldr_locale]
   end
 
   test "locale is set according to the configured priority" do
-    opts = Cldr.Plug.SetLocale.init(from: [:accept_language, :query], cldr: TestBackend.Cldr)
+    opts = Cldr.Plug.SetLocale.init(from: [:accept_language, :query], cldr_backend: TestBackend.Cldr)
 
     conn =
       :get
@@ -224,6 +224,6 @@ defmodule Cldr.Plug.SetLocale.Test do
                language_variant: nil
              }
 
-    assert opts[:cldr].get_current_locale() == conn.private[:cldr_locale]
+    assert opts[:cldr_backend].get_current_locale() == conn.private[:cldr_locale]
   end
 end